requirements.txt
# ============================================
# QuantTrade - Complete Requirements
# Python 3.10+
# ============================================

# Core Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Data Models & Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Environment & Configuration
python-dotenv>=1.0.0

# Data Processing & Analysis
pandas>=2.1.3
numpy>=1.26.2
scikit-learn>=1.3.2

# Machine Learning
catboost>=1.2.2
joblib>=1.3.2

# Visualization
matplotlib>=3.8.2

# Telegram Bot
python-telegram-bot>=20.7
requests>=2.31.0

# Database (Optional - for future use)
sqlalchemy>=2.0.23
psycopg2-binary>=2.9.9

# HTTP Client
httpx>=0.25.2

# CORS
fastapi-cors>=0.0.6

# Utilities
pytz>=2023.3
